requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
attrs==23.1.0
certifi==2022.6.15.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-taggit==4.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
numpy==1.24.1
oauthlib==3.2.1
Pillow==9.2.0
proglog==0.1.10
psycopg2==2.9.3
pytz==2023.3
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
Send2Trash==1.8.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.5.0
tzdata==2022.2
urllib3==1.26.12

